==> lake_problem/notebooks/02. generate reference sets.ipynb <==
import pandas as pd
import pareto
import os
import sys
from itertools import chain
from matplotlib import pyplot as plt

sys.path.append("..")

import rbf_functions

# Load solutions for each RBF for all seeds:

import os

rbfs = [
    rbf_functions.squared_exponential_rbf,
    rbf_functions.original_rbf,
    rbf_functions.inverse_quadratic_rbf,
    rbf_functions.inverse_multiquadric_rbf,
    rbf_functions.exponential_rbf,
    rbf_functions.matern32_rbf,
    rbf_functions.matern52_rbf,
]

root_dir = "../output//"
pareto_sets = {}
for entry in rbfs:
    sollist = []
    solutions = []
    name = entry.__name__
    output_dir = os.path.join(root_dir, name)

    for filename in os.listdir(output_dir):
        if filename.endswith("solution.csv"):
            sollist.append(filename[:-4])
            df_temp = pd.read_csv(f"{output_dir}/{filename}", header=0)
            solutions.append(df_temp.values.tolist())
    pareto_sets[name] = list(chain.from_iterable(solutions))

pareto_sets.keys()
print("Amount of solutions for each RBF:")
for rbf in pareto_sets:
    print(rbf, len(pareto_sets[rbf]))
# print(f"Loaded into list 'solutions': {', '.join(sollist)}")

# Create reference set for each RBF, and save:

reference_sets = {}
for rbf in pareto_sets:
    print(rbf, len(pareto_sets[rbf]))
    nondominated = pareto.eps_sort(
        [pareto_sets[rbf]], [0, 1, 2, 3], [0.05, 0.05, 0.05, 0.05], maximize=[1, 2, 3]
    )
    reference_sets[rbf] = nondominated
    df_nondom = pd.DataFrame(
        nondominated, columns=["max_p", "utility", "inertia", "reliability"]
    )
    print(rbf, len(df_nondom))
    df_nondom.to_csv(f"./{rbf}_refset.csv", index=False, header=True)

# Find decision variables that belong to the generated refset:

d_refvar = {}
d_refsol = {}
for entry in rbfs:
    name = entry.__name__
    output_dir = os.path.join(root_dir, name)

    ts = []
    tv = []
    for filename in os.listdir(output_dir):
        # load
        if filename.endswith("solution.csv"):
            df_ts = pd.read_csv(f"{output_dir}/{filename}", header=0)
            ts.append(df_ts)
        elif filename.endswith("variables.csv"):
            df_tv = pd.read_csv(f"{output_dir}/{filename}", header=None)
            tv.append(df_tv)

    df_sol = pd.concat(ts)
    df_var = pd.concat(tv)

    sol = df_sol.values.tolist()
    var = df_var.values.tolist()
    ref = reference_sets[name]
    refvar = []
    refsol = []
    for idx, value in enumerate(sol):
        if value in ref:
            refvar.append(var[idx])
            refsol.append(sol[idx])
    d_refsol[name] = pd.DataFrame(
        refsol, columns=["max_p", "utility", "inertia", "reliability"]
    )
    d_refvar[name] = pd.DataFrame(refvar)
    d_refsol[name].to_csv(f"./refsets/{name}_refset.csv", index=False, header=True)
    d_refvar[name].to_csv(
        f"./refsets/{name}_refset_variables.csv", index=False, header=False
    )

for entry in rbfs:
    name = entry.__name__
    print(name)
    print(f"refset: {len(d_refsol[name])}")
    print(f"varset: {len(d_refvar[name])}")

# Generate global reference set for all RBFs:

x = 0
for rbf in pareto_sets:
    x += len(pareto_sets[rbf])
    print(rbf, len(pareto_sets[rbf]))
print("total:", x)

pareto_set = {}
sollist = []
solutions = []
for entry in rbfs:
    name = entry.__name__
    output_dir = os.path.join(root_dir, name)
    for filename in os.listdir(output_dir):
        if filename.endswith("solution.csv"):
            sollist.append(filename[:-4])
            df_temp = pd.read_csv(f"{output_dir}/{filename}", header=0)
            solutions.append(df_temp.values.tolist())
pareto_set = list(chain.from_iterable(solutions))
len(pareto_set)

print(len(pareto_set))
nondominated = pareto.eps_sort(
    [pareto_set], [0, 1, 2, 3], [0.05, 0.05, 0.05, 0.05], maximize=[1, 2, 3]
)
df_nondom = pd.DataFrame(
    nondominated, columns=["max_p", "utility", "inertia", "reliability"]
)
print(len(nondominated))
df_nondom.to_csv(f"./refsets/all_rbf_refset.csv", index=False, header=True)
