==> lake_problem/notebooks/14. MannWhitney analysis.ipynb <==
# When comparing two data sets using statistical tools, it is not
# sufficient to simply compare their average or median values. This is because
# such results can be skewed by randomness. For example, suppose we are
# calculating the median values of ten seeds. If one algorithm gets "lucky"
# and happens to use more above-average seeds, the estimated median will
# be skewed. Therefore, it is necessary to check the statistical significance of
# results. This is exactly what the indifferent entries are displaying. To determine
# statistical significance, the Kruskal-Wallis and Mann-Whitney U tests with 95% confidence
# intervals can be used.

# The Mann-Whitney U test is a nonparametric test of the null hypothesis that the distribution
# underlying sample x is the same as the distribution underlying sample y.
# It is often used as a test of difference in location between distributions.

from scipy import stats
import sys
import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt

plt.rcParams["figure.figsize"] = [12, 8]
sys.path.append("..")
import rbf_functions

rbfs = [
    rbf_functions.original_rbf,
    rbf_functions.squared_exponential_rbf,
    rbf_functions.inverse_quadratic_rbf,
    rbf_functions.inverse_multiquadric_rbf,
    rbf_functions.exponential_rbf,
    rbf_functions.matern32_rbf,
    rbf_functions.matern52_rbf,
]

# Mann-Whitney HV local refset

gddata = {}
eidata = {}
hvdata = {}
varlist = ["10", "20", "30", "40", "50", "60", "70", "80", "90", "100"]
for entry in rbfs:
    rbf = entry.__name__
    #     df_gd_data = pd.read_csv(f"gd_all/{rbf}_gd_all.csv")
    #     df_ei_data = pd.read_csv(f"ei_all/{rbf}_ei_all.csv")
    df_hv_data = pd.read_csv(f"hv/{rbf}_hv.csv")
    gddata[rbf] = {}
    eidata[rbf] = {}
    hvdata[rbf] = {}
    for i in varlist:  # add max hv value at 100.000 nfe to dict for each seed
        #         gddata[rbf][i] = df_gd_data[i].dropna().iloc[-1]
        #         eidata[rbf][i] = df_ei_data[i].dropna().iloc[-1]
        hvdata[rbf][i] = df_hv_data[i].dropna().iloc[-1]
# df_gd = pd.DataFrame.from_dict(gddata, orient='index')
# df_ei = pd.DataFrame.from_dict(eidata, orient='index')
df_hv = pd.DataFrame.from_dict(hvdata, orient="index")

# Max attained hypervolume per rbf for each seed
df_hv

# method:
# 'asymptotic': compares the standardized test statistic against the normal distribution, correcting for ties.
# 'exact': computes the exact p-value by comparing the observed  statistic against the exact distribution of the  statistic under the null hypothesis. No correction is made for ties.
# 'auto': chooses 'exact' when the size of one of the samples is less than 8 and there are no ties; chooses 'asymptotic' otherwise.

df_mannwhitney = pd.DataFrame(columns=df_hv.index)
df_pval = pd.DataFrame(columns=df_hv.index)
for i, row in df_hv.iterrows():
    for j, row2 in df_hv.iterrows():
        mannwhit, pval = stats.mannwhitneyu(row, row2, method="auto")
        df_mannwhitney.loc[i, j] = mannwhit
        df_pval.loc[i, j] = pval

df_mannwhitney

df_pval

# Mann-Whitney  on objectives

rbfd = {}
rbfn = []
for rbf in rbfs:
    name = rbf.__name__
    rbfn.append(name)
    rbfd[name] = pd.read_csv(f"refsets/{name}_refset.csv")

# The lower the U is, the more different the groups are.
# https://www.youtube.com/watch?v=nRAAAp1Bgnw
# https://www.youtube.com/watch?v=Twk6lBhBl88

# Hydropower

df_mannwhit_hyd = pd.DataFrame(columns=rbfn)
df_pval_hyd = pd.DataFrame(columns=rbfn)
for i in rbfd:
    for j in rbfd:
        row1 = rbfd[i]["hydropower"]
        row2 = rbfd[j]["hydropower"]
        mannwhit, pval = stats.mannwhitneyu(row1, row2, method="auto")
        df_mannwhit_hyd.loc[i, j] = mannwhit
        df_pval_hyd.loc[i, j] = pval

df_mannwhit_hyd

df_pval_hyd

# Atomic power plant

df_mannwhit_atom = pd.DataFrame(columns=rbfn)
df_pval_atom = pd.DataFrame(columns=rbfn)
for i in rbfd:
    for j in rbfd:
        row1 = rbfd[i]["atomicpowerplant"]
        row2 = rbfd[j]["atomicpowerplant"]
        mannwhit, pval = stats.mannwhitneyu(row1, row2, method="auto")
        df_mannwhit_atom.loc[i, j] = mannwhit
        df_pval_atom.loc[i, j] = pval

df_mannwhit_atom

df_pval_atom

# Baltimore

df_mannwhit_balt = pd.DataFrame(columns=rbfn)
df_pval_balt = pd.DataFrame(columns=rbfn)
for i in rbfd:
    for j in rbfd:
        row1 = rbfd[i]["baltimore"]
        row2 = rbfd[j]["baltimore"]
        mannwhit, pval = stats.mannwhitneyu(row1, row2, method="auto")
        df_mannwhit_balt.loc[i, j] = mannwhit
        df_pval_balt.loc[i, j] = pval

df_mannwhit_balt

df_pval_balt

# Chester

df_mannwhit_ches = pd.DataFrame(columns=rbfn)
df_pval_ches = pd.DataFrame(columns=rbfn)
for i in rbfd:
    for j in rbfd:
        row1 = rbfd[i]["chester"]
        row2 = rbfd[j]["chester"]
        mannwhit, pval = stats.mannwhitneyu(row1, row2, method="auto")
        df_mannwhit_ches.loc[i, j] = mannwhit
        df_pval_ches.loc[i, j] = pval

df_mannwhit_ches

df_pval_ches

# Environment

df_mannwhit_env = pd.DataFrame(columns=rbfn)
df_pval_env = pd.DataFrame(columns=rbfn)
for i in rbfd:
    for j in rbfd:
        row1 = rbfd[i]["environment"]
        row2 = rbfd[j]["environment"]
        mannwhit, pval = stats.mannwhitneyu(row1, row2, method="auto")
        df_mannwhit_env.loc[i, j] = mannwhit
        df_pval_env.loc[i, j] = pval

df_mannwhit_env

df_pval_env

# Recreation

df_mannwhit_rec = pd.DataFrame(columns=rbfn)
df_pval_rec = pd.DataFrame(columns=rbfn)
for i in rbfd:
    for j in rbfd:
        row1 = rbfd[i]["recreation"]
        row2 = rbfd[j]["recreation"]
        mannwhit, pval = stats.mannwhitneyu(row1, row2, method="auto")
        df_mannwhit_rec.loc[i, j] = mannwhit
        df_pval_rec.loc[i, j] = pval

df_mannwhit_rec

df_pval_rec

# create index
df = df_mannwhit_hyd.astype(float).round(2)
newcol = df.columns.to_list()
newcol = [x[:-4] for x in newcol]
df.columns = newcol
df.index = newcol

df = df_mannwhit_rec.astype(float).round(1)
df2 = df_pval_rec.astype(float).round(2)
df.columns = newcol
df.index = newcol
df2.columns = newcol
df2.index = newcol
print("\\begin{table}[h]")
# print("\\parbox{.45\linewidth}{")
print("\\setlength{\\tabcolsep}{3.5pt}")
print("\\begin{tabular}{c | " + " | ".join(["c"] * len(df.columns)) + "}")
# print(" & ".join([str(x) for x in newcol]) + " \\\\")
print(
    "& original & \makecell{squared\\\exponential} & \makecell{inverse\\\quadratic} & \makecell{inverse\\\multiquadric} & exponential & matern32 & matern52 \\\\"
)
print("\hline")
for i, row in df.iterrows():
    j = i.replace("_", " ")
    print(f"{j} & " + " & ".join([str(x) for x in row.values]) + " \\\\")
print("\\end{tabular}")
print("\\caption{Mann-Whitney test, recreation}")
print("\\end{table}")
print("")
# print("\\parbox{.45\linewidth}{")
# print("\\centering")
print("\\begin{table}[h]")
print("\\begin{tabular}{c | " + " | ".join(["c"] * len(df.columns)) + "}")
# print(" & " + " & ".join([str(x) for x in newcol]))
print(
    "& original & \makecell{squared\\\exponential} & \makecell{inverse\\\quadratic} & \makecell{inverse\\\multiquadric} & exponential & matern32 & matern52 \\\\"
)
print("\hline")
for i, row in df2.iterrows():
    j = i.replace("_", " ")
    print(f"{j} & " + " & ".join([str(x) for x in row.values]) + " \\\\")
print("\\end{tabular}")
print("\\caption{Mann-Whitney test, P value, recreation}")
# print("}")
print("\\end{table}")
